# web/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Artemis Studio API",
    "description" : "Cluster-wide management and observability for Apache ActiveMQ Artemis",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/clusters/{clusterId}/queues" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "queues",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewQueueView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "createQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateQueueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/queues/{queueName}/resume" : {
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "resumeQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/queues/{queueName}/reset-counter" : {
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "resetCounter",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/queues/{queueName}/pause" : {
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "pauseQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/addresses" : {
      "get" : {
        "tags" : [ "cluster-resource-controller" ],
        "operationId" : "addresses",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ResourceQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedViewAddressView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "createAddress",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/queues/{queueName}" : {
      "delete" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "deleteQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "override",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "updateQueue",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateQueueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClusterDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/addresses/{address}" : {
      "delete" : {
        "tags" : [ "queue-lifecycle-controller" ],
        "operationId" : "deleteAddress",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "address",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifecycleOutcomeView"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "CapabilitiesView" : {
        "type" : "object",
        "properties" : {
          "managementRead" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "managementWrite" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "notifications" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "messageIo" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "slowConsumerDetection" : {
            "$ref" : "#/components/schemas/CapabilityView"
          }
        },
        "required" : [ "managementRead", "managementWrite", "messageIo", "notifications", "slowConsumerDetection" ]
      },
      "CapabilityView" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "brokerXmlSnippet" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "reason", "status" ]
      },
...
      "LogicalNodeView" : {
        "type" : "object",
        "properties" : {
          "artemisNodeId" : {
            "type" : [ "string", "null" ]
          },
          "splitBrain" : {
            "type" : "string"
          },
          "replicationBehind" : {
            "type" : "boolean"
          },
          "endpoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NodeEndpointView"
            }
          }
        },
        "required" : [ "endpoints", "replicationBehind", "splitBrain" ]
      },
      "NodeEndpointView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "artemisNodeId" : {
            "type" : [ "string", "null" ]
          },
          "jolokiaUrl" : {
            "type" : [ "string", "null" ]
          },
          "coreUrl" : {
            "type" : [ "string", "null" ]
          },
          "haRole" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "replicaSync" : {
            "type" : [ "boolean", "null" ]
          },
          "version" : {
            "type" : [ "string", "null" ]
          },
          "lastError" : {
            "type" : [ "string", "null" ]
          },
          "lastSeenAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "discovered" : {
            "type" : "boolean"
          },
          "manualOverride" : {
            "type" : "boolean"
          },
          "manageable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "active", "discovered", "haRole", "id", "manageable", "manualOverride", "name", "state" ]
      },
...
      "TopologyView" : {
        "type" : "object",
        "properties" : {
          "clusterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LogicalNodeView"
            }
          }
        },
        "required" : [ "clusterId", "nodes" ]
      },
      "ClusterDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "topology" : {
            "$ref" : "#/components/schemas/TopologyView"
          },
          "capabilities" : {
            "$ref" : "#/components/schemas/CapabilitiesView"
          },
          "health" : {
            "$ref" : "#/components/schemas/HealthView"
          },
          "environmentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        },
        "required" : [ "capabilities", "health", "id", "name", "topology" ]
      },
      "HealthView" : {
        "type" : "object",
        "properties" : {
          "clusterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "level" : {
            "type" : "string"
          },
          "liveEndpointNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "splitBrain" : {
            "type" : "string"
          },
          "replicationBehind" : {
            "type" : "boolean"
          },
          "notes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "clusterId", "level", "liveEndpointNames", "notes", "replicationBehind", "splitBrain" ]
      },
...
      "CreateQueueRequest" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string",
            "description" : "The address this queue binds to.",
            "minLength" : 1
          },
          "name" : {
            "type" : "string",
            "description" : "The queue's name.",
            "minLength" : 1
          },
          "routingType" : {
            "type" : "string",
            "description" : "ANYCAST or MULTICAST. Cannot be changed once the queue exists.",
            "enum" : [ "ANYCAST", "MULTICAST" ],
            "minLength" : 1,
            "pattern" : "(?i)ANYCAST|MULTICAST"
          },
          "durable" : {
            "type" : "boolean",
            "default" : true,
            "description" : "Whether the queue survives a broker restart."
          },
          "filter" : {
            "type" : [ "string", "null" ],
            "description" : "A JMS selector limiting what the queue accepts."
          },
          "maxConsumers" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Maximum concurrent consumers; -1 for unlimited."
          },
          "purgeOnNoConsumers" : {
            "type" : [ "boolean", "null" ],
            "description" : "Delete the queue when its last consumer disconnects."
          },
          "exclusive" : {
            "type" : [ "boolean", "null" ],
            "description" : "Route to exactly one consumer at a time."
          },
          "nonDestructive" : {
            "type" : [ "boolean", "null" ],
            "description" : "Messages stay on the queue after delivery."
          },
          "ringSize" : {
            "type" : [ "integer", "null" ],
            "format" : "int64",
            "description" : "Retain at most this many messages; -1 for unlimited."
          },
          "autoCreateAddress" : {
            "type" : "boolean",
            "default" : true,
            "description" : "Create the address if it does not exist."
          }
        },
        "required" : [ "address", "name", "routingType" ]
      },
      "LifecycleOutcomeView" : {
        "type" : "object",
        "properties" : {
          "dryRun" : {
            "type" : "boolean",
            "description" : "Whether this was a preview that mutated nothing."
          },
          "cap" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "The bulk safety cap in force."
          },
          "overCap" : {
            "type" : "boolean",
            "description" : "Whether the summed estimate exceeds the cap."
          },
          "partial" : {
            "type" : "boolean",
            "description" : "Whether some nodes applied the command and others did not."
          },
          "totalAffected" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total messages destroyed, or that would be."
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NodeOutcomeView"
            }
          }
        },
        "required" : [ "cap", "dryRun", "nodes", "overCap", "partial", "totalAffected" ]
      },
      "NodeOutcomeView" : {
        "type" : "object",
        "properties" : {
          "nodeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodeName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "description" : "WOULD_APPLY on a preview; APPLIED when the node was changed; ALREADY when it was already in the requested state; SKIPPED_NOT_LIVE when the node was not live and never received the command; FAILED when it refused.",
            "enum" : [ "WOULD_APPLY", "APPLIED", "ALREADY", "SKIPPED_NOT_LIVE", "FAILED" ]
          },
          "affected" : {
            "type" : [ "integer", "null" ],
            "format" : "int64",
            "description" : "Messages destroyed, or that would be destroyed, on this node."
          },
          "error" : {
            "type" : [ "string", "null" ],
            "description" : "Why this node failed."
          }
        },
        "required" : [ "nodeId", "nodeName", "status" ]
      },
...
      "CreateAddressRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "The address name.",
            "minLength" : 1
          },
          "routingTypes" : {
            "type" : "string",
            "description" : "Comma-separated routing types the address supports.",
            "enum" : [ "ANYCAST", "MULTICAST", "ANYCAST,MULTICAST" ],
            "minLength" : 1,
            "pattern" : "(?i)ANYCAST|MULTICAST|ANYCAST,MULTICAST|MULTICAST,ANYCAST"
          }
        },
        "required" : [ "name", "routingTypes" ]
      },
...
      "UpdateQueueRequest" : {
        "type" : "object",
        "properties" : {
          "filter" : {
            "type" : [ "string", "null" ],
            "description" : "A JMS selector limiting what the queue accepts."
          },
          "maxConsumers" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "purgeOnNoConsumers" : {
            "type" : [ "boolean", "null" ]
          },
          "exclusive" : {
            "type" : [ "boolean", "null" ]
          },
          "nonDestructive" : {
            "type" : [ "boolean", "null" ]
          },
          "ringSize" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        }
      },
...
      "ResourceQuery" : {
        "type" : "object",
        "properties" : {
          "q" : {
            "type" : "string"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : "string"
          }
        }
      },
...
      "PagedViewQueueView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QueueView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
      "QueueNodeCell" : {
        "type" : "object",
        "properties" : {
          "nodeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodeName" : {
            "type" : "string"
          },
          "stale" : {
            "type" : "boolean"
          },
          "lastSeenAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "messageCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "consumerCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deliveringCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "scheduledCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paused" : {
            "type" : "boolean"
          }
        },
        "required" : [ "consumerCount", "deliveringCount", "messageCount", "nodeId", "nodeName", "paused", "scheduledCount", "stale" ]
      },
      "QueueView" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "queueName" : {
            "type" : "string"
          },
          "routingType" : {
            "type" : "string"
          },
          "durable" : {
            "type" : "boolean"
          },
          "totalMessageCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalConsumerCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalDeliveringCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalScheduledCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nodesPresent" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nodesTotal" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paused" : {
            "type" : "boolean",
            "description" : "True when the queue is paused on at least one node. A queue paused on some nodes and not others is a divergence the operator needs to see, so this is deliberately 'any', not 'all' — perNode says which."
          },
          "perNode" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QueueNodeCell"
            }
          }
        },
        "required" : [ "address", "durable", "nodesPresent", "nodesTotal", "paused", "perNode", "queueName", "routingType", "totalConsumerCount", "totalDeliveringCount", "totalMessageCount", "totalScheduledCount" ]
      },
...
      "AddressView" : {
        "type" : "object",
        "properties" : {
          "nodeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodeName" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "routingTypes" : {
            "type" : [ "string", "null" ]
          },
          "queueCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "messageCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "messageCount", "name", "nodeId", "nodeName", "queueCount" ]
      },
      "PagedViewAddressView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AddressView"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "count", "data", "page", "pageSize" ]
      },
...
    }
  }
}